feat: Added Project DELETE Authentication

+++ BugReporterBe/src/projects/providers/deleteProject.provider.ts
@@ -1,18 +1,22 @@
 import type { Request, Response } from "express";
 import { StatusCodes, ReasonPhrases } from "http-status-codes";
 import Project from "../projects.schema.ts";
+import Bug from "../../bugs/bugs.schema.ts";
 import { matchedData } from "express-validator";
 import errorLogger from "../../helpers/errorLogger.helper.ts";
 
 async function deleteProjectProvider(req: Request, res: Response)
 {
   const validatedResult = matchedData(req); 
+  const id = validatedResult.projectId;
 
   try {
+    const project = await Project.findOne({ _id: id, user: req.user?.sub });
+    if (!project) return res.status(StatusCodes.NOT_FOUND).json({ reason: "No Project found for the provided ID" });
+    
+    await Bug.deleteMany({ project: id });
 
-    const id = validatedResult.projectId;
     const deletedProject = await Project.deleteOne({ _id: id });
-
     return res.status(StatusCodes.OK).json(deletedProject);
   }
   catch (error) {
